fix(typing): add pyright config and harden runtime-safe type narrowing
- configure pyright in pyproject with project include/exclude rules and execution environments
- add pyright to dev dependencies
- tighten null/metadata/state guards across detector, propagation, signal, simulator, plotter, and platform modules
- add protocol-based narrowing for dynamic interfaces while preserving duck-typed runtime behaviour
- keep public APIs unchanged and resolve current pyright diagnostics

+++ bluepebble/sigproc/beamformer.py
@@ -585,7 +585,7 @@ class MinimumVarianceDistortionlessResponseBeamformer(Beamformer):
         if pad > 0:
             x_array = np.pad(x_array, ((0, 0), (0, pad)), mode="constant")
 
-        window = np.hanning(nfft).astype(x_array.real.dtype)
+        window = np.hanning(nfft).astype(np.float64)
         # Make a 3D view: (M, n_frames, nfft)
         stride_t = x_array.strides[1]
         frames = np.lib.stride_tricks.as_strided(
